+++ src/textual/command.py
@@ -555,7 +555,7 @@ class CommandPalette(SystemModalScreen[None]):
     CommandPalette > .command-palette--help-text {                   
         color: $text-muted;
         background: transparent;
-        text-style: not bold dim;       
+        text-style: not bold;       
     }
     
     CommandPalette > .command-palette--highlight {
@@ -571,6 +571,7 @@ class CommandPalette(SystemModalScreen[None]):
         height: 100%;
         visibility: hidden;
         background: $surface;
+        &:dark { background: $panel-darken-1; }
     }
 
     CommandPalette #--input {